Moved actions out to its own component.
