+++ webapp/src/app/resource/components/actions/actions.component.scss
@@ -1,0 +1,23 @@
+@import "colors";
+
+.share-popover {
+    h6 {
+        margin: 0;
+        color: $brand-green;
+    }
+    .copy-clipboard {
+        width: 100%;
+        display:inline-flex;
+        align-items: center;
+        justify-items: center;
+        justify-content: space-between;
+        input {
+            font-size: 1rem;
+            flex-grow: 1;
+        }
+    }
+    .copy-success {
+        font-size: 10px;
+        margin-left: -6px;
+    }
+}
