+++ webapp/src/app/resource/components/actions/actions.component.spec.ts
@@ -1,0 +1,30 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ActionsComponent } from './actions.component';
+import { ResourceModule } from '../../resource.module';
+import { mockDataServiceProviders } from 'src/app/app.module.spec';
+import { mockResourceModel } from 'src/app/data/mock-data';
+
+describe('ActionsComponent', () => {
+  let component: ActionsComponent;
+  let fixture: ComponentFixture<ActionsComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [ ResourceModule ],
+      providers: mockDataServiceProviders
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(ActionsComponent);
+    component = fixture.componentInstance;
+    component.model = mockResourceModel;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
