+++ webapp/src/app/resource/instructional/content/resource-content.component.html
@@ -1,32 +1,9 @@
 <div class="top-header">
   <img src="/assets/images/yellow-spot.png" />
-  <div class="actions">
-    <sbdl-button-icon (click)="share()"><i class="far fa-share"></i></sbdl-button-icon>
-    <ng-container #shareButton></ng-container>
-    <sbdl-button-icon (click)="toggleFavorite()">
-      <div *ngIf="!togglingFavorite && model.details.favorite">
-        <i class="fas fa-bookmark"></i>
-      </div>
-      <div *ngIf="!togglingFavorite && !model.details.favorite">
-        <i class="far fa-bookmark"></i>
-      </div>
-      <div *ngIf="togglingFavorite">
-        <i class="fas fa-spinner fa-spin"></i>
-      </div>
-    </sbdl-button-icon>
-    <sbdl-button-icon><i class="far fa-print"></i></sbdl-button-icon>
-    <sbdl-button-icon [hidden]="hideReadingModeToggle" (click)="toggleReadingMode()">
-      <div *ngIf="readingMode">
-        <i class="far fa-compress-alt"></i>
-      </div>
-      <div *ngIf="!readingMode">
-        <i class="far fa-expand-alt"></i>
-      </div>
-    </sbdl-button-icon>
-  </div>
+  <sbdl-actions [model]= "model" (readingModeChanged)="emitReadingModeChanged($event)"></sbdl-actions>
 </div>
 <header>
-  <h1 #getStarted>{{ details.title }}</h1>
+  <h1>{{ details.title }}</h1>
   <section class="byline">
     {{ details.author }}&nbsp;&nbsp;&bull;&nbsp;&nbsp;<span
       class="font-light"
@@ -50,15 +27,4 @@
 <sbdl-differentiation [model]="model.differentiation" (sectionElementLoaded)="setDifferentiation($event)"></sbdl-differentiation>
 <hr/>
 <sbdl-formative [model]="model.formative" (sectionElementLoaded)="setFormative($event)"></sbdl-formative>
-<ng-template #sharePopover>
-  <div class="share-popover">
-      <h6>Share this resource</h6>
-      <div class="copy-clipboard">
-          <input #url readonly type="text" [value]="currentUrl">
-          <sbdl-button-icon (click)="copyToClipboard(url)">
-            <i class="fas fa-copy" title="Copy to clipboard"></i>
-            <div class="copy-success" [hidden]="!showCopied">Copied!</div>
-          </sbdl-button-icon>
-      </div>
-  </div>
-</ng-template>
+
