+++ webapp/src/app/resource/resource.module.ts
@@ -16,6 +16,7 @@ import { ProfessionalResourceComponent } from './professional/professional-resou
 import { ResourceHostDirective } from './resource-host.directive';
 import { ResourceTypeStrategyComponent } from './resource-type-strategy.component';
 import { ResourceComponent } from './resource.component';
+import { ActionsComponent } from './components/actions/actions.component';
 
 export function getResourceComponents(resourceComponent: ResourceComponent) {
   return resourceComponent;
@@ -34,7 +35,7 @@ export function getResourceComponents(resourceComponent: ResourceComponent) {
     FormativeComponent, 
     StepByStepComponent, 
     AttachmentsComponent, 
-    ProfessionalResourceComponent
+    ProfessionalResourceComponent, ActionsComponent
   ],
   imports: [
     CommonModule,
